+++ src/main/java/roomescape/controller/ThemeController.java
@@ -1,48 +1,19 @@
 package roomescape.controller;
 
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-import roomescape.domain.theme.ThemeRequest;
-import roomescape.domain.theme.ThemeResponse;
-import roomescape.service.ThemeService;
 
-import java.net.URI;
-import java.util.List;
-
-@RestController
+@Controller
 public class ThemeController {
 
-    private final ThemeService themeService;
-
-    public ThemeController(ThemeService themeService) {
-        this.themeService = themeService;
+    @GetMapping("/theme")
+    public String themePage() {
+        return "theme";
     }
 
-    @PostMapping("/admin/themes")
-    public ResponseEntity<ThemeResponse> create(@RequestBody ThemeRequest themeRequest) {
-        ThemeResponse newTheme = themeService.create(themeRequest);
-        return ResponseEntity.created(URI.create("/admin/themes/" + newTheme.getId())).body(newTheme);
-    }
-
-    @GetMapping("/themes")
-    public ResponseEntity<List<ThemeResponse>> readAll() {
-        return ResponseEntity.ok().body(themeService.findAll());
-    }
-
-    @GetMapping("/themes/popular")
-    public ResponseEntity<List<ThemeResponse>> readPopularTheme() {
-        return ResponseEntity.ok().body(themeService.findPopularTheme());
-    }
-
-    @DeleteMapping("/admin/themes/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Long id) {
-        themeService.delete(id);
-        return ResponseEntity.noContent().build();
+    @GetMapping("/admin/theme")
+    public String adminThemePage() {
+        return "admin/theme";
     }
 }
