+++ src/main/resources/application.properties
@@ -4,3 +4,6 @@ spring.datasource.url=jdbc:h2:mem:database
 spring.datasource.driver-class-name=org.h2.Driver
 spring.datasource.username=sa
 spring.datasource.password=
+
+spring.mvc.view.prefix=/WEB-INF/jsp/
+spring.mvc.view.suffix=.jsp
