+++ src/test/java/roomescape/admin/MissionStep4Test.java
@@ -5,7 +5,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.controller.ReservationController;
+import roomescape.controller.ReservationRestController;
 
 import java.lang.reflect.Field;
 
@@ -16,7 +16,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 public class MissionStep4Test {
 
     @Autowired
-    private ReservationController reservationController;
+    private ReservationRestController reservationController;
 
     @Test
     void 계층화_리팩터링() {
